Code Review
HTML is the root element and it should always contain the lang attribute set to your primary language. if lang is not set. by default it will be set to unknown. meta charset was also missing. UTF-8 encoding is super important & it should be the first element beneath the head. here lang attribute is set to Spanish(es). I hope your language is Spanish.